# docs/swagger.yaml
basePath: /
definitions:
...
  entites.Engresso:
    properties:
      cpf:
        type: string
      data_nascimento:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Esta é a documentação da API dos microserviços.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Microservices API
  version: "1.0"
paths:
  /categories/createCategory:
    post:
      consumes:
      - application/json
      description: para realizar o cadastros do engresso precisa name,cpf,data nascimento.
      parameters:
      - description: Dados ebresso
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/entites.Engresso'
      produces:
      - application/json
      responses: {}
      summary: create engresso
      tags:
      - auth
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"